Cleaning up, commenting
