+++ tools/ci-build/publisher/src/subcommand/tag_versions_manifest.rs
@@ -5,8 +5,8 @@
 
 use anyhow::{bail, Result};
 use clap::Parser;
-use smithy_rs_tool_common::versions_manifest::{Release, VersionsManifest};
-use std::{collections::BTreeMap, path::PathBuf};
+use smithy_rs_tool_common::versions_manifest::VersionsManifest;
+use std::path::PathBuf;
 
 #[derive(Parser, Debug)]
 pub struct TagVersionsManifestArgs {
@@ -23,15 +23,10 @@ pub fn subcommand_tag_versions_manifest(
 ) -> Result<()> {
     println!("Tagging manifest at: {manifest_path:#?}");
     let mut manifest = VersionsManifest::from_file(manifest_path)?;
-    let mut release = manifest.release.as_mut();
-    if let Some(rel) = release {
-        rel.tag = Some(tag.to_string());
+    if let Some(release) = manifest.release.as_mut() {
+        release.tag = Some(tag.to_string());
     } else {
-        release = Some(&mut Release {
-            tag: Some(tag.to_string()),
-            crates: BTreeMap::new(),
-        });
-        // bail!("The given versions manifest file doesn't have a `[release]` section in it to add the tag to");
+        bail!("The given versions manifest file doesn't have a `[release]` section in it to add the tag to");
     }
     manifest.write_to_file(manifest_path)?;
     Ok(())
